Optimize imports

+++ src/main/java/tfm/programs/pdg/Example1.java
@@ -33,7 +33,7 @@ public class Example1 {
             y *= x;
         }
 
-        int e = (Integer) x;
+        int e = x;
 
         switch (x) {
             case 1:
